Just cosmetic changes

+++ grbl/SpindleControl.c
@@ -27,14 +27,15 @@
 #include "GCode.h"
 #include "SpindleControl.h"
 #include "Config.h"
+#include "stm32f4xx_tim.h"
 
 
-static float pwm_gradient; // Precalulated value to speed up rpm to PWM conversions.
+static float pwm_gradient; // Precalculated value to speed up rpm to PWM conversions.
 static uint8_t spindle_enabled = 0;
 
 
-void Spindle_Init(void)
-{
+void
+Spindle_Init(void) {
     // Configure variable spindle PWM and enable pin, if requried. On the Uno, PWM and enable are
     // combined unless configured otherwise.
     GPIO_InitGPIO(GPIO_SPINDLE);
@@ -50,8 +51,8 @@ void Spindle_Init(void)
 // Disables the spindle and sets PWM output to zero when PWM variable spindle speed is enabled.
 // Called by various main program and ISR routines. Keep routine small, fast, and efficient.
 // Called by spindle_init(), spindle_set_speed(), spindle_set_state(), and mc_reset().
-void Spindle_Stop(void)
-{
+void
+Spindle_Stop(void) {
 	TIM1->CCR1 = 100; // Disable PWM. Output voltage is zero.
 	spindle_enabled = 0;
 
@@ -63,11 +64,10 @@ void Spindle_Stop(void)
 }
 
 
-uint8_t Spindle_GetState(void)
-{
+uint8_t
+Spindle_GetState(void) {
     // Check if PWM is enabled.
-	if(spindle_enabled)
-    {
+    if(spindle_enabled) {
 		/*if(SPINDLE_DIRECTION_PORT & (1<<SPINDLE_DIRECTION_BIT)) {
 			return SPINDLE_STATE_CCW;
 		}
@@ -83,8 +83,8 @@ uint8_t Spindle_GetState(void)
 
 // Sets spindle speed PWM output and enable pin, if configured. Called by spindle_set_state()
 // and stepper ISR. Keep routine small and efficient.
-void Spindle_SetSpeed(uint8_t pwm_value)
-{
+void
+Spindle_SetSpeed(uint8_t pwm_value) {
 	TIM1->CCR1 = 100 - pwm_value; // Set PWM output level.
 #ifdef SPINDLE_ENABLE_OFF_WITH_ZERO_SPEED
 	if (pwm_value == SPINDLE_PWM_OFF_VALUE) {
@@ -117,7 +117,7 @@ uint8_t Spindle_ComputePwmValue(float rpm) // 328p PWM register is 8-bit.
 {
 	uint8_t pwm_value;
 
-	rpm *= (0.010*sys.spindle_speed_ovr); // Scale by spindle speed override value.
+    rpm *= (0.010f*sys.spindle_speed_ovr); // Scale by spindle speed override value.
 
 	// Calculate PWM register value based on rpm max/min settings and programmed rpm.
 	if((settings.rpm_min >= settings.rpm_max) || (rpm >= settings.rpm_max)) {
@@ -126,7 +126,7 @@ uint8_t Spindle_ComputePwmValue(float rpm) // 328p PWM register is 8-bit.
 		pwm_value = SPINDLE_PWM_MAX_VALUE;
 	}
 	else if(rpm <= settings.rpm_min) {
-		if(rpm == 0.0) { // S0 disables spindle
+        if(rpm == 0.0f) { // S0 disables spindle
 			sys.spindle_speed = 0.0;
 			pwm_value = SPINDLE_PWM_OFF_VALUE;
 		}
@@ -139,7 +139,7 @@ uint8_t Spindle_ComputePwmValue(float rpm) // 328p PWM register is 8-bit.
 		// Compute intermediate PWM value with linear spindle speed model.
 		// NOTE: A nonlinear model could be installed here, if required, but keep it VERY light-weight.
 		sys.spindle_speed = rpm;
-		pwm_value = floor((rpm-settings.rpm_min)*pwm_gradient) + SPINDLE_PWM_MIN_VALUE;
+        pwm_value = (uint8_t)(floorf((rpm-settings.rpm_min)*pwm_gradient)) + SPINDLE_PWM_MIN_VALUE;
 	}
 
 	return pwm_value;
@@ -149,8 +149,8 @@ uint8_t Spindle_ComputePwmValue(float rpm) // 328p PWM register is 8-bit.
 // Immediately sets spindle running state with direction and spindle rpm via PWM, if enabled.
 // Called by g-code parser spindle_sync(), parking retract and restore, g-code program end,
 // sleep, and spindle stop override.
-void Spindle_SetState(uint8_t state, float rpm)
-{
+void
+Spindle_SetState(uint8_t state, float rpm) {
 	if(sys.abort) {
 		// Block during abort.
 		return;
@@ -185,8 +185,8 @@ void Spindle_SetState(uint8_t state, float rpm)
 
 // G-code parser entry-point for setting spindle state. Forces a planner buffer sync and bails
 // if an abort or check-mode is active.
-void Spindle_Sync(uint8_t state, float rpm)
-{
+void
+Spindle_Sync(uint8_t state, float rpm) {
 	if(sys.state == STATE_CHECK_MODE) {
 		return;
 	}
